Add two extra arguments for llvm.dbg.value
In case of salvaging a binary instruction which has two registers
as operands, the first argument of the llvm.dbg.value will be set
to the first of the two registers, the fourth argument will
be set to the second register and the fifth argument will be set
to an empty DIExpression.

When salvaging the instruction that assigns a value to the register
which is the fourth argument in some llvm.dbg.value call, the
DIExpression that is produced will be added to the fifth argument
of the same llvm.dbg.value, and the fourth argument will be
replaced with the register that is used in that instruction.

Also add new internal DWARF operations in that case, which will
use the fourth argument (operations like DW_OP_LLVM_reg_plus,
DW_OP_LLVM_reg_minus).

Signed-off-by: Mateja Marjanovic <Mateja.Marjanovic@rt-rk.com>