Add an extra register and DIExpression for the DBG_VALUE in the backend
The IR passed llvm.dbg.value with five arguments, the backend
made two extra arguments for DBG_VALUE using the fourth and
fifth arguments (an extra register and an extra DIExpression)
of the llvm.dbg.value.

Those two combined make the value of the second operand in
an instruction in which both arguments are registers.

In case of an instruction in which one operand is a register, and
the other is an immediate value, there is no need for an extra
register or DIExpression, so $noreg and DIExpression() are the
values of those extra arguments.

Signed-off-by: Mateja Marjanovic <Mateja.Marjanovic@rt-rk.com>