Add implementation of new internal DWARF operations in AsmPrinter
When salvaging in the IR, in case of an instruction in which both
arguments are registers, we add new DW_OP_LLVM_reg_* operations in
the DIExpression of the llvm.dbg.value call for the described
variable.
In case of salvaging the instruction of adding two registers
(%a and %b) and assigning the value to the third register
(%c), DW_OP_LLVM_reg_plus will be added to the DIExpression
of the llvm.dbg.value in which %c was the first argument.

DW_OP_LLVM_reg_* operations are implemented by finding the
DWARF register number of the fifth argument of the DBG_VALUE
(the register that is used in describing the second operand
of the instruction in which the operands are two registers)
and then replacing the DW_OP_LLVM_reg_* operation in the
DIExpression of the current DBG_VALUE with the following:
- a DW_OP_breg* (DWARF register number, as mentioned above)
- everything from the second DIExpression (sixth argument),
  which combined with the DW_OP_breg* describes the value of
  the second operand in the instruction with two registers
- a proper DW_OP_* operation (for DW_OP_LLVM_reg_plus, the
  proper operation would be DW_OP_plus, etc.)

Signed-off-by: Mateja Marjanovic <Mateja.Marjanovic@rt-rk.com>