[testsuite] Add support for new llvm.dbg.value/DBG_VALUE instructions
Since llvm intrinsic call llvm.dbg.value and machine instruction
DBG_VALUE have two new arguments (one representing the second register
and the other the second DIExpression), tests have to be modified
so they could be passed. We don't have to modify the code of the test,
but the CHECK part of it. The code of the test will be processed in
llvm/lib/IR/AutoUpgrade.cpp, but we need to add the default values
for the new arguments (metadata undef, metadata !DIExpression() for
llvm.dbg.value and $noreg, !DIExpression() for the DBG_VALUE).

Signed-off-by: Mateja Marjanovic <Mateja.Marjanovic@rt-rk.com>